requirements.txt
Django
djangorestframework
django-cors-headers==4.3.1
Pillow==10.1.0
opencv-python==4.8.1.78
mediapipe==0.10.7
tensorflow==2.14.0
numpy==1.24.3
scikit-learn==1.3.2
pandas==1.5.3
python-decouple==3.8
celery==5.3.4
redis==5.0.1
gunicorn==21.2.0
psycopg2-binary==2.9.9
django-allauth==0.57.0
django-crispy-forms==2.1
crispy-bootstrap5==0.7
channels==4.0.0
channels-redis==4.1.0
whitenoise==6.6.0
django-extensions==3.2.3
reportlab==4.0.7